+++ src/pages/index/main.vue
@@ -1,9 +1,19 @@
 <template>
   <div class="index-page">
-    <carrousel :articles="articleListStore.list.data" :fetching="articleListStore.list.fetching" />
-    <announcement
-      :announcements="announcementStore.announcements"
-      :fetching="announcementStore.fetching || articleListStore.list.fetching"
+    <carrousel
+      class="carrousel"
+      :articles="articleListStore.list.data"
+      :fetching="articleListStore.list.fetching"
+    />
+    <twitter
+      class="twitter"
+      :userinfo="twitterStore.userinfo.data || void 0"
+      :tweets="twitterStore.tweets.data || void 0"
+      :fetching="
+        twitterStore.userinfo.fetching ||
+        twitterStore.tweets.fetching ||
+        articleListStore.list.fetching
+      "
     />
     <article-list
       :mammon="false"
@@ -18,28 +28,29 @@
 <script lang="ts">
   import { defineComponent } from 'vue'
   import { useEnhancer } from '/@/app/enhancer'
+  import { isSSR } from '/@/app/environment'
   import { useUniversalFetch, onClient } from '/@/universal'
-  import { useAnnouncementStore } from '/@/store/announcement'
+  import { useTwitterStore } from '/@/store/twitter'
   import { useArticleListStore } from '/@/store/article'
   import { useMetaStore } from '/@/store/meta'
   import { nextScreen } from '/@/utils/effects'
   import { META } from '/@/config/app.config'
-  import Carrousel from '/@/components/flow-desktop/carrousel.vue'
-  import Announcement from '/@/components/flow-desktop/announcement.vue'
   import ArticleList from '/@/components/flow-desktop/list.vue'
+  import Carrousel from './carrousel.vue'
+  import Twitter from './twitter.vue'
 
   export default defineComponent({
     name: 'IndexPage',
     components: {
       Carrousel,
-      Announcement,
+      Twitter,
       ArticleList
     },
     setup() {
       const { meta } = useEnhancer()
       const metaStore = useMetaStore()
+      const twitterStore = useTwitterStore()
       const articleListStore = useArticleListStore()
-      const announcementStore = useAnnouncementStore()
 
       meta(() => ({
         title: `${META.title} - ${META.sub_title}`,
@@ -55,15 +66,32 @@
         }
       }
 
-      useUniversalFetch(() =>
-        Promise.all([articleListStore.fetchList(), announcementStore.fetchList()])
-      )
+      useUniversalFetch(() => {
+        const twitterUserinfoPromise = twitterStore.fetchUserinfo()
+        const twitterTweetsPromise = twitterStore.fetchTweets()
+        const promises = [articleListStore.fetchList()]
+        if (isSSR) {
+          promises.push(twitterUserinfoPromise, twitterTweetsPromise)
+        }
+        return Promise.all(promises)
+      })
 
       return {
+        twitterStore,
         articleListStore,
-        announcementStore,
         loadmoreArticles
       }
     }
   })
 </script>
+
+<style lang="scss" scoped>
+  @import 'src/styles/init.scss';
+
+  .index-page {
+    .carrousel,
+    .twitter {
+      margin-bottom: $lg-gap;
+    }
+  }
+</style>